aoc-solutions/src/test/java/org/thermoweb/aoc/days/Day6Test.java
package org.thermoweb.aoc.days;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.math.BigInteger;
import java.util.Optional;

import org.junit.jupiter.api.Test;
import org.thermoweb.aoc.Day;
import org.thermoweb.aoc.DayRunner;

class Day6Test {
    private final Day day = new Day6();

    @Test
    void test_part_one() throws Exception {
        assertEquals(Optional.of(BigInteger.valueOf(41)), day.partOne(DayRunner.getExample(6)));
    }

    @Test
    void test_part_two() throws Exception {
        assertEquals(Optional.of(BigInteger.valueOf(6)), day.partTwo(DayRunner.getExample(6)));
    }
}
